# perf(sql): use last_course_visit mv for learner summary query