Performance improvements for learner summary

```
# Initial Version

5. individual-learner - Learner Summary
Superset time: 2.011835 (s).
Query duration: 0.17 (s).
Result rows: 100
Memory Usage (MB): 30.506019592285156
Row count (superset) 100
Filters: [{'column': 'emission_time'}]

# Changes for last_visit MV

11. individual-learner - Learner Summary
Superset time: 0.915487 (s).
Query duration: 0.047 (s).
Result rows: 100
Memory Usage (MB): 12.710393905639648
Row count (superset) 100
Filters: [{'column': 'emission_time'}]
```

Performance improvements for Learner Performance:

```
# Before
0 rows in set. Elapsed: 2.204 sec. Processed 3.84 million rows, 440.55 MB (1.74 million rows/s., 199.91 MB/s.)
Peak memory usage: 758.23 MiB.

# After
0 rows in set. Elapsed: 0.598 sec. Processed 3.84 million rows, 440.55 MB (6.42 million rows/s., 736.68 MB/s.)
Peak memory usage: 133.32 MiB.
```